# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UNILIV Admin API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      parameters:
        - in: query
          name: propertyId
          schema:
            type: string
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStatsResponse"
  /dashboard/charts:
    get:
      operationId: getDashboardCharts
      tags: [dashboard]
      summary: Get dashboard chart data
      parameters:
        - in: query
          name: propertyId
          schema:
            type: string
      responses:
        "200":
          description: Dashboard charts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardChartsResponse"

  # ─── PROPERTIES ───────────────────────────────────────
...
  /complaints:
    get:
      operationId: getComplaints
      tags: [complaints]
      summary: List complaints
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
        - $ref: "#/components/parameters/PropertyIdParam"
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: priority
          schema:
            type: string
      responses:
        "200":
          description: Complaints list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintsListResponse"
    post:
      operationId: createComplaint
      tags: [complaints]
      summary: Create complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComplaintBody"
      responses:
        "201":
          description: Complaint created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintResponse"
  /complaints/{id}:
    get:
      operationId: getComplaint
      tags: [complaints]
      summary: Get complaint by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Complaint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintResponse"
    put:
      operationId: updateComplaint
      tags: [complaints]
      summary: Update complaint
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComplaintBody"
      responses:
        "200":
          description: Complaint updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintResponse"
  /escalations:
    post:
      operationId: createEscalation
      tags: [complaints]
      summary: Create escalation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEscalationBody"
      responses:
        "201":
          description: Escalation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscalationResponse"

  # ─── EMPLOYEES ────────────────────────────────────────
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
    SearchParam:
      name: search
      in: query
      schema:
        type: string
    PropertyIdParam:
      name: propertyId
      in: query
      schema:
        type: string
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [total, page, limit, totalPages]

    # ─── AUTH ─────────────────────────────────────────────
...
    DashboardStats:
      type: object
      properties:
        totalProperties:
          type: integer
        totalResidents:
          type: integer
        totalBeds:
          type: integer
        occupiedBeds:
          type: integer
        occupancyRate:
          type: number
        openComplaints:
          type: integer
        criticalComplaints:
          type: integer
        totalEmployees:
          type: integer
        pendingLeaves:
          type: integer
        newLeadsThisMonth:
          type: integer
        convertedLeadsThisMonth:
          type: integer
        revenueThisMonth:
          type: number
        pendingPayments:
          type: number
        lowStockItems:
          type: integer
      required: [totalProperties, totalResidents, totalBeds, occupiedBeds, occupancyRate, openComplaints, criticalComplaints, totalEmployees, pendingLeaves, newLeadsThisMonth, convertedLeadsThisMonth, revenueThisMonth, pendingPayments, lowStockItems]
    DashboardStatsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/DashboardStats"
      required: [success, data]
    ChartDataPoint:
      type: object
      properties:
        label:
          type: string
        value:
          type: number
      required: [label, value]
    DashboardCharts:
      type: object
      properties:
        occupancyTrend:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        complaintsByCategory:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        revenueTrend:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        leadsByStage:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        attendanceThisMonth:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
      required: [occupancyTrend, complaintsByCategory, revenueTrend, leadsByStage, attendanceThisMonth]
    DashboardChartsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/DashboardCharts"
      required: [success, data]

    # ─── PROPERTY ─────────────────────────────────────────
...
    ComplaintDto:
      type: object
      properties:
        id:
          type: string
        propertyId:
          type: string
        residentId:
          type: string
          nullable: true
        ticketNo:
          type: string
        category:
          type: string
        subCategory:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        assignedTo:
          type: string
          nullable: true
        slaHours:
          type: integer
        slaDeadline:
          type: string
          nullable: true
        slaBreach:
          type: boolean
        resolvedAt:
          type: string
          nullable: true
        rating:
          type: integer
          nullable: true
        residentName:
          type: string
          nullable: true
        propertyName:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, propertyId, ticketNo, category, title, description, status, priority, slaHours, slaBreach, createdAt]
    ComplaintResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/ComplaintDto"
      required: [success, data]
    ComplaintsListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/ComplaintDto"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [success, data, meta]
    CreateComplaintBody:
      type: object
      properties:
        propertyId:
          type: string
        residentId:
          type: string
        category:
          type: string
        subCategory:
          type: string
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        assignedTo:
          type: string
        slaHours:
          type: integer
      required: [propertyId, category, title, description]
    UpdateComplaintBody:
      type: object
      properties:
        status:
          type: string
        priority:
          type: string
        assignedTo:
          type: string
        resolutionNote:
          type: string
        rating:
          type: integer
    EscalationDto:
      type: object
      properties:
        id:
          type: string
        complaintId:
          type: string
        level:
          type: integer
        escalatedTo:
          type: string
        reason:
          type: string
        createdAt:
          type: string
      required: [id, complaintId, level, escalatedTo, reason, createdAt]
    EscalationResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/EscalationDto"
      required: [success, data]
    CreateEscalationBody:
      type: object
      properties:
        complaintId:
          type: string
        level:
          type: integer
        escalatedTo:
          type: string
        reason:
          type: string
      required: [complaintId, level, escalatedTo, reason]

    # ─── EMPLOYEE ─────────────────────────────────────────
...